add docker build of container
it would be nice to have a quick demo container
or Dockerfile to build one for quickly testing out
the software without needing to install anything
to my system. This addition will add the Dockerfile,
instructions in the README for build and usage,
along with an automated build to deploy a container
alongside the repository.

Signed-off-by: vsoch <vsoch@users.noreply.github.com>